[#42] Remove overlapping compounds
We call remove listener, and clean the cells. If compounds overlap, we
might call listener with the cells which were already cleaned. If
listener callback tries to access any fields (e.g. getName() by
logger), if crashed the UI with NullException, and it looked like the
second compound wasn't removed (though it actually was removed).

This fix ensures that field is cleaned only after all listeners are
fired.